[EXPERIMENTAL] Removed old architecture
